requirements.txt
# Function dependencies, for example:
# package>=version
# Function dependencies, for example:
# package>=version
google-cloud-storage==1.42.3
gcsfs==2021.10.1
Cython==0.29.24
numpy==1.21.2
pandas==1.3.4
google-cloud-logging==1.15.0
croniter==1.3.5
jupyterlab
tqdm
python-dotenv=1.0.0